% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projB-package.R
\docType{package}
\name{projB-package}
\alias{projB}
\alias{projB-package}
\title{projB: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ramnath Vaidyanathan \email{ramnath.vaidya@gmail.com}

Authors:
\itemize{
  \item Ramnath Vaidyanathan \email{ramnath.vaidya@gmail.com}
}

}
\keyword{internal}
